requirements-base.txt
# requirements-base.txt
# 
# Базові залежності проєкту без ML (Machine Learning) бібліотек.
# Цей файл використовується для швидкого встановлення залежностей у CI/CD,
# особливо для unit тестів, які не потребують ML функціональності.
#
# Для повної функціональності (включно з ML) також встановіть requirements-ml.txt:
#   pip install -r requirements-base.txt -r requirements-ml.txt
#
# Або використовуйте оригінальний requirements.txt для локальної розробки.
#
# ============================================================================
# Core Framework & API
# ============================================================================
fastapi==0.111.0
fastapi-cli==0.0.13
starlette==0.37.2
uvicorn==0.30.0
pydantic==2.8.2
pydantic_core==2.20.1
annotated-types==0.7.0

# ============================================================================
# Database & ORM
# ============================================================================
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
greenlet==3.2.4

# ============================================================================
# Authentication & Security
# ============================================================================
bcrypt==5.0.0
passlib==1.7.4
python-jose==3.5.0
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.1

# ============================================================================
# HTTP & Networking
# ============================================================================
httpx==0.28.1
httpcore==1.0.9
httptools==0.6.4
h11==0.16.0
websockets==15.0.1
anyio==4.11.0
sniffio==1.3.1
idna==3.10
certifi==2025.8.3
urllib3==2.5.0
dnspython==2.8.0

# ============================================================================
# Data Processing & Analysis (без ML, тільки базові)
# ============================================================================
numpy==2.3.3
pandas==2.3.2
scikit-learn==1.7.2  # Використовується для TF-IDF fallback в paraphrase.py
scipy==1.16.2
joblib==1.5.2
threadpoolctl==3.6.0

# ============================================================================
# Cloud Services
# ============================================================================
cloudinary==1.44.1

# ============================================================================
# Validation & Parsing
# ============================================================================
email-validator==2.3.0
python-multipart==0.0.20
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# ============================================================================
# Serialization & Formatting
# ============================================================================
orjson==3.11.3
ujson==5.11.0
PyYAML==6.0.1
markdown-it-py==4.0.0
mdurl==0.1.2

# ============================================================================
# Templating & Rendering
# ============================================================================
Jinja2==3.1.6
MarkupSafe==3.0.3

# ============================================================================
# CLI & Utilities
# ============================================================================
click==8.3.0
typer==0.19.2
rich==14.1.0
rich-toolkit==0.15.1
Pygments==2.19.2
colorama==0.4.6
shellingham==1.5.4

# ============================================================================
# Configuration & Environment
# ============================================================================
python-dotenv==1.1.1
iniconfig==2.1.0
packaging==25.0

# ============================================================================
# Testing
# ============================================================================
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
coverage==7.11.0

# ============================================================================
# Scheduling & Background Tasks
# ============================================================================
APScheduler==3.10.4

# ============================================================================
# Development & Build Tools
# ============================================================================
watchfiles==1.1.0
pluggy==1.6.0
six==1.17.0
typing_extensions==4.15.0

